requirements.txt
django
gunicorn
django-heroku
djangorestframework
requests
environ